requirements.txt
sqlalchemy
flask
flask-sqlalchemy
flask-cors
psycopg2
gunicorn
requests
simplejson
flask-login
google-auth
